+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 * [CONRAT-27817](https://jira.rakuten-it.com/jira/browse/CONRAT-27817): Remove dependency on RSDKUtils.
 * [CONRAT-27901](https://jira.rakuten-it.com/jira/browse/CONRAT-27901): Create a documentation for Internal JSON Serialization in ADVANCED_USAGE.md
 
+### Sonarqube fixes
+* [CONRAT-28044](https://jira.rakuten-it.com/jira/browse/CONRAT-28044): Fix Reachability.swift SonarQube issues.
+
 ### Unit tests
 * [CONRAT-27983](https://jira.rakuten-it.com/jira/browse/CONRAT-27983): Improve the RAT Tracking tests with CoreHelpers injection.